+++ Lab8/FixedVector.h
@@ -15,7 +15,7 @@ namespace lab8
 		bool Remove(const T& t);
 		const T& Get(const size_t index);
 		T& operator[](const size_t index);
-		size_t GetIndex(const T& t);
+		int GetIndex(const T& t);
 		size_t GetSize();
 		size_t GetCapacity();
 
@@ -82,7 +82,7 @@ namespace lab8
 	}
 
 	template <typename T, size_t N>
-	size_t FixedVector<T, N>::GetIndex(const T& t)
+	int FixedVector<T, N>::GetIndex(const T& t)
 	{
 		for (size_t i = 0; i < mSize; ++i)
 		{
@@ -91,7 +91,7 @@ namespace lab8
 				continue;
 			}
 
-			return i;
+			return static_cast<int>(i);
 		}
 
 		return -1;
